+++ PerseusMeteo/MeteoDataAPIs/OpenWeather/OpenWeatherClient.swift
@@ -1,5 +1,5 @@
 //
-//  OpenWeatherStar.swift
+//  OpenWeatherClient.swift
 //  Version: 0.3.5
 //
 //  Created by Mikhail Zhigulin in 7531.